(ref) reorganize utility functions for server
