--- service/src/server/messages.js
@@ -1,9 +1,0 @@
-import dotenv from 'dotenv';
-
-import service from './index';
-
-dotenv.config();
-
-// const pollSqs = () => {
-  
-// }
